Run tests on PHPUnit 9
