+++ composer.json
@@ -25,6 +25,6 @@
     "require-dev": {
         "clue/buzz-react": "^2.4",
         "clue/ndjson-react": "^1.0",
-        "phpunit/phpunit": "^7.0 || ^6.4 || ^5.7 || ^4.8.35"
+        "phpunit/phpunit": "^9.3 || ^5.7 || ^4.8.35"
     }
 }
