+++ phpunit.xml.dist
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<phpunit bootstrap="vendor/autoload.php" colors="true">
+<phpunit bootstrap="vendor/autoload.php" colors="true" cacheResult="false">
     <testsuites>
         <testsuite name="Flux test suite">
             <directory>./tests/</directory>
