+++ tests/TransformerAllTest.php
@@ -3,7 +3,6 @@
 namespace Clue\Tests\React\Flux;
 
 use Clue\React\Flux\Transformer;
-use PHPUnit\Framework\TestCase;
 use React\Promise\Deferred;
 use React\Promise\Promise;
 use React\Stream\ThroughStream;
@@ -202,43 +201,5 @@ class TransformerAllTest extends TestCase
         $first->reject(new \RuntimeException());
 
         $promise->then(null, $this->expectCallableOnce());
-    }
-
-    protected function expectCallableOnce()
-    {
-        $mock = $this->createCallableMock();
-
-        $mock
-            ->expects($this->once())
-            ->method('__invoke');
-
-        return $mock;
-    }
-
-    protected function expectCallableOnceWith($param)
-    {
-        $mock = $this->createCallableMock();
-
-        $mock
-            ->expects($this->once())
-            ->method('__invoke')
-            ->with($param);
-
-        return $mock;
-    }
-
-    protected function expectCallableNever()
-    {
-        $mock = $this->createCallableMock();
-        $mock
-            ->expects($this->never())
-            ->method('__invoke');
-
-        return $mock;
-    }
-
-    protected function createCallableMock()
-    {
-        return $this->getMockBuilder('stdClass')->setMethods(array('__invoke'))->getMock();
     }
 }
