+++ tests/TransformerAnyTest.php
@@ -3,7 +3,6 @@
 namespace Clue\Tests\React\Flux;
 
 use Clue\React\Flux\Transformer;
-use PHPUnit\Framework\TestCase;
 use React\Promise\Deferred;
 use React\Promise\Promise;
 use React\Stream\ThroughStream;
@@ -210,43 +209,5 @@ class TransformerAnyTest extends TestCase
         $first->resolve('hello');
 
         $promise->then($this->expectCallableOnceWith('hello'));
-    }
-
-    protected function expectCallableOnce()
-    {
-        $mock = $this->createCallableMock();
-
-        $mock
-            ->expects($this->once())
-            ->method('__invoke');
-
-        return $mock;
-    }
-
-    protected function expectCallableOnceWith($param)
-    {
-        $mock = $this->createCallableMock();
-
-        $mock
-            ->expects($this->once())
-            ->method('__invoke')
-            ->with($param);
-
-        return $mock;
-    }
-
-    protected function expectCallableNever()
-    {
-        $mock = $this->createCallableMock();
-        $mock
-            ->expects($this->never())
-            ->method('__invoke');
-
-        return $mock;
-    }
-
-    protected function createCallableMock()
-    {
-        return $this->getMockBuilder('stdClass')->setMethods(array('__invoke'))->getMock();
     }
 }
